+++ src/features/home/contexts/types.ts
@@ -1,5 +1,0 @@
-import { EpisodeFragment } from '../../../generated/graphql';
-
-export type EpisodeType = EpisodeFragment & {
-  loading?: boolean;
-};