+++ src/features/home/features/upcoming/components/Upcoming.tsx
@@ -1,0 +1,10 @@
+import UpcomingProvider from '../contexts/UpcomingContext';
+import UpcomingSection from './UpcomingSection';
+
+const Upcoming = () => (
+  <UpcomingProvider>
+    <UpcomingSection />
+  </UpcomingProvider>
+);
+
+export default Upcoming;