+++ src/features/home/features/upnext/components/UpNext.tsx
@@ -1,0 +1,10 @@
+import UpNextProvider from '../contexts/UpNextContext';
+import UpNextSection from './UpNextSection';
+
+const UpNext = () => (
+  <UpNextProvider>
+    <UpNextSection />
+  </UpNextProvider>
+);
+
+export default UpNext;