+++ src/features/shows/features/episode/components/UpsertEpisodeAction.tsx
@@ -2,7 +2,7 @@ import CheckCircleIcon from '@mui/icons-material/CheckCircle';
 import CheckCircleOutlineIcon from '@mui/icons-material/CheckCircleOutline';
 import { useCallback, useContext } from 'react';
 import ActionButton from '../../../components/base/ActionButton';
-import { UpNextContext } from '../../../../home/contexts/UpNextContext';
+import { UpNextContext } from '../../../../home/features/upnext/contexts/UpNextContext';
 import { ActionProps } from '../types';
 
 const UpsertEpisodeAction = ({ episode }: ActionProps) => {