+++ src/features/shows/features/episode/types.ts
@@ -1,4 +1,8 @@
-import type { EpisodeType } from '../../../home/contexts/types';
+import { EpisodeFragment } from '../../../../generated/graphql';
+
+export type EpisodeType = EpisodeFragment & {
+  loading?: boolean;
+};
 
 export interface ActionProps {
   episode: EpisodeType;